Define separate implementation for desktop app
