+++ src/libs/actions/Welcome/index.ts
@@ -5,7 +5,6 @@ import * as API from '@libs/API';
 import {SIDE_EFFECT_REQUEST_COMMANDS, WRITE_COMMANDS} from '@libs/API/types';
 import DateUtils from '@libs/DateUtils';
 import Log from '@libs/Log';
-import CONST from '@src/CONST';
 import type {OnboardingCompanySize} from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
 import type {OnboardingPurpose} from '@src/types/onyx';
@@ -13,6 +12,7 @@ import type Onboarding from '@src/types/onyx/Onboarding';
 import type TryNewDot from '@src/types/onyx/TryNewDot';
 import {isEmptyObject} from '@src/types/utils/EmptyObject';
 import {clearInitialPath} from './OnboardingFlow';
+import switchToOldDotOnNonMicroCompanySize from './switchToOldDotOnNonMicroCompanySize';
 
 type OnboardingData = Onboarding | undefined;
 
@@ -148,29 +148,6 @@ function completeHybridAppOnboarding() {
         Log.info(`[HybridApp] Onboarding status has changed. Propagating new value to OldDot`, true);
         NativeModules.HybridAppModule.completeOnboarding(true);
     });
-}
-
-/*
- * Decides whether we should set dismissed key to `true` or `false` based on company size
- */
-function switchToOldDotOnNonMicroCompanySize(onboardingCompanySize: OnboardingCompanySize) {
-    if (onboardingCompanySize === CONST.ONBOARDING_COMPANY_SIZE.MICRO) {
-        return;
-    }
-
-    const optimisticData: OnyxUpdate[] = [
-        {
-            onyxMethod: Onyx.METHOD.MERGE,
-            key: ONYXKEYS.NVP_TRYNEWDOT,
-            value: {
-                classicRedirect: {
-                    dismissed: true,
-                },
-            },
-        },
-    ];
-
-    API.write(WRITE_COMMANDS.SWITCH_TO_OLD_DOT_ON_COMPANY_SIZE, {onboardingCompanySize}, {optimisticData});
 }
 
 Onyx.connect({
