+++ src/libs/actions/Welcome/switchToOldDotOnNonMicroCompanySize/index.desktop.ts
@@ -1,0 +1,6 @@
+import type SwitchToOldDotOnNonMicroCompanySize from './types';
+
+// We don't want to use this API on desktop, so we just noop here
+const switchToOldDotOnNonMicroCompanySize: SwitchToOldDotOnNonMicroCompanySize = () => {};
+
+export default switchToOldDotOnNonMicroCompanySize;
