+++ src/libs/actions/Welcome/switchToOldDotOnNonMicroCompanySize/types.ts
@@ -1,0 +1,5 @@
+import type {OnboardingCompanySize} from '@src/CONST';
+
+type SwitchToOldDotOnNonMicroCompanySize = (onboardingCompanySize: OnboardingCompanySize) => void;
+
+export default SwitchToOldDotOnNonMicroCompanySize;
